+++ scraper/rimi/src/main/java/GroceryFamily/GroceryDad/rimi/RimiScraper.java
@@ -1,4 +1,4 @@
-package GroceryFamily.GroceryDad;
+package GroceryFamily.GroceryDad.rimi;
 
 import com.codeborne.selenide.SelenideElement;
 import lombok.Builder;
@@ -11,64 +11,60 @@ import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 import java.math.BigDecimal;
-import java.net.URLDecoder;
-import java.nio.charset.StandardCharsets;
 import java.util.ArrayList;
 import java.util.Collection;
 
 import static com.codeborne.selenide.Condition.text;
 import static com.codeborne.selenide.Condition.visible;
 import static com.codeborne.selenide.Selenide.*;
-import static org.apache.commons.lang3.StringUtils.substringAfterLast;
 
 @Slf4j
 @SpringBootApplication
-public class PrismaScraper implements CommandLineRunner {
+public class RimiScraper implements CommandLineRunner {
     public static void main(String... args) {
-        SpringApplication.run(PrismaScraper.class, args);
+        SpringApplication.run(RimiScraper.class, args);
     }
 
     final WebDriver driver;
 
-    PrismaScraper(WebDriver driver) {
+    RimiScraper(WebDriver driver) {
         this.driver = driver;
     }
 
     @Override
     public void run(String... args) {
         using(driver, () -> {
-            open("https://prismamarket.ee");
-            switchToEnglish();
-            closeCookieNotice();
-            category("Groceries", "Pets", "Toys for pets");
+            open("https://rimi.ee/epood/en");
+            useOnlyStrictlyNecessaryCookies();
+            category("Groceries", "All category products");
             var products = products();
             log.info("Products: {}", products);
             // todo: implement further
         });
     }
 
-    static void switchToEnglish() {
-        $("*[data-language='en']").shouldBe(visible).click();
-    }
-
-    static void closeCookieNotice() {
-        $("*[class*='js-cookie-notice'] *[class='close-icon']").shouldBe(visible).click();
+    static void useOnlyStrictlyNecessaryCookies() {
+        $("#CybotCookiebotDialogBodyLevelButtonLevelOptinDeclineAll").shouldBe(visible).click();
     }
 
     static void category(String... names) {
         if (names.length == 0) return;
-        $$("*[class='main-navigation-items'] a").findBy(text(names[0])).click();
-        if (names.length == 1) return;
-        $$("*[class*='left-navigation'] a").findBy(text(names[1])).click();
-        if (names.length == 2) return;
-        $$("*[class*='categories-shelf'] a").findBy(text(names[2])).click();
-        if (names.length > 3) throw new IllegalArgumentException("Too many categories");
+        $("#desktop_category_menu_button").shouldBe(visible).click();
+        var name = names[0];
+        var button = $$("#desktop_category_menu button").findBy(text(name));
+        var submenu = button.attr("aria-owns");
+        button.click();
+        for (int idx = 1; idx < names.length; ++idx) {
+            name = names[idx];
+            button = $$("#" + submenu + " a").findBy(text(name));
+            submenu = button.attr("aria-owns");
+            button.click();
+        }
     }
 
     static Collection<Product> products() {
-        $("*[class*='js-products-shelf']").shouldBe(visible);
         Collection<Product> products = new ArrayList<>();
-        for (var e : $$("*[class*='js-shelf-item']")) {
+        for (var e : $$("*[class='product-grid__item'] > div")) {
             products.add(new Product(e));
         }
         return products;
@@ -80,23 +76,13 @@ public class PrismaScraper implements CommandLineRunner {
         final String code;
         final String name;
         final Price pcPrice;
+        final Price kgPrice;
 
         Product(SelenideElement e) {
-            code = code(e);
-            name = e.$("*[class='name']").text();
-            pcPrice = Price.pc(e.$("*[class*='js-comp-price']").text());
-        }
-
-        static String code(SelenideElement e) {
-            return decode(substringAfterLast(url(e), "/"));
-        }
-
-        static String url(SelenideElement e) {
-            return e.$("a").attr("href");
-        }
-
-        static String decode(String url) {
-            return URLDecoder.decode(url, StandardCharsets.UTF_8);
+            code = e.attr("data-product-code");
+            name = e.$("*[class='card__name']").text();
+            pcPrice = Price.pc(e.$("*[class*='price-tag']").text());
+            kgPrice = Price.kg(e.$("*[class='card__price-per']").text());
         }
     }
 
@@ -108,15 +94,26 @@ public class PrismaScraper implements CommandLineRunner {
         final BigDecimal value;
         final String currency;
 
-        // 6,99 €/pcs
+        // 4\n29\n€/pcs.
         static Price pc(String text) {
+            var fragments = text.split("\n");
+            return Price
+                    .builder()
+                    .type("pc")
+                    .value(new BigDecimal(fragments[0] + '.' + fragments[1]))
+                    .currency(fragments[2].substring(0, 1))
+                    .build();
+        }
+
+        // 8,09 € /kg
+        static Price kg(String text) {
             var fragments = text.split(" ");
             var value = fragments[0].split(",");
             return Price
                     .builder()
-                    .type("pc")
+                    .type("kg")
                     .value(new BigDecimal(value[0] + '.' + value[1]))
-                    .currency(fragments[1].substring(0, 1))
+                    .currency(fragments[1])
                     .build();
         }
     }
